LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-grid/subgrid/subgrid-baseline-005-expected.txt








PASS .item 1
FAIL .item 2 assert_equals:
<div data-offset-y="85" class="item last">
        <span></span><br><span></span>
      </div>
offsetTop expected 85 but got 93
FAIL .item 3 assert_equals:
<div data-offset-y="163" class="item first">
        <span></span><br><span></span>
      </div>
offsetTop expected 163 but got 171
FAIL .item 4 assert_equals:
<div data-offset-y="218" class="item last">
        <span></span><br><span></span>
      </div>
offsetTop expected 218 but got 233
FAIL .item 5 assert_equals:
<div data-offset-y="308" class="item first">
      <span></span><br><span></span>
    </div>
offsetTop expected 308 but got 313
FAIL .item 6 assert_equals:
<div data-offset-y="360" class="item last">
      <span></span><br><span></span>
    </div>
offsetTop expected 360 but got 375
FAIL .item 7 assert_equals:
<div data-offset-y="41" class="item small first"></div>
offsetTop expected 41 but got 3
FAIL .item 8 assert_equals:
<div data-offset-y="110" class="item small last"></div>
offsetTop expected 110 but got 125
FAIL .item 9 assert_equals:
<div data-offset-y="168" class="item small first"></div>
offsetTop expected 168 but got 153
FAIL .item 10 assert_equals:
<div data-offset-y="243" class="item small last"></div>
offsetTop expected 243 but got 275
FAIL .item 11 assert_equals:
<div data-offset-y="313" class="item small first"></div>
offsetTop expected 313 but got 303
FAIL .item 12 assert_equals:
<div data-offset-y="385" class="item small last"></div>
offsetTop expected 385 but got 425
